requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
twilio
python-multipart
requests
pymongo
rapidfuzz
openai-whisper
edge-tts
aiofiles
certifi